+++ test/action_decoder_test.dart
@@ -1,13 +1,13 @@
 import 'package:test/test.dart';
-import '../lib/redux_remote_devtools.dart';
+import 'package:redux_remote_devtools/redux_remote_devtools.dart';
 
 void main() {
   group('NOPActionDecoder', () {
     group('decode', () {
       test('Passes through the json payload', () {
         var payload = {'type': 'SOME ACTION', 'value': 123};
-        var decoder = new NopActionDecoder();
-        var result = decoder.decode(payload);
+        var decoder = NopActionDecoder;
+        var result = decoder(payload);
         expect(result, payload);
       });
     });