+++ test/state_encoder_test.dart
@@ -1,12 +1,12 @@
 import 'dart:convert';
 import 'package:test/test.dart';
-import '../lib/redux_remote_devtools.dart';
+import 'package:redux_remote_devtools/redux_remote_devtools.dart';
 
 class TestState {
   int value;
   TestState({this.value});
-  toJson() {
-    return {'value': this.value};
+  Map<String, dynamic> toJson() {
+    return {'value': value};
   }
 }
 
@@ -21,15 +21,15 @@ void main() {
   group('JsonStateEncoder', () {
     group('encode', () {
       test('Returns a jsonified state', () {
-        var encoder = new JsonStateEncoder();
-        var result = encoder.encode(new TestState(value: 5));
+        var encoder = JsonStateEncoder;
+        var result = encoder(TestState(value: 5));
         var decoded = jsonDecode(result);
         expect(decoded['value'], equals(5));
       });
       test('Throws an exception if unencodable', () {
-        var encoder = new JsonStateEncoder();
+        var encoder = JsonStateEncoder;
         var testFunc = () {
-          encoder.encode(new TestUnencodableState(value: 5));
+          encoder(TestUnencodableState(value: 5));
         };
         expect(testFunc, throwsA(TypeMatcher<JsonUnsupportedObjectError>()));
       });